Phonebook - Separate components into modules
