+++ Part2/phonebook/src/components/EntryForm.js
@@ -1,0 +1,19 @@
+const EntryForm = ({onSubmit, nameValue, numberValue, nameChange, numberChange}) => {
+    return(
+      <form onSubmit={onSubmit}>
+        <div>
+          name:
+          <input value={nameValue} onChange={nameChange} />
+        </div>
+        <div>
+          number:
+          <input value={numberValue} onChange={numberChange} />
+        </div>
+        <div>
+          <button type="submit">add</button>
+        </div>
+      </form>
+    )
+  }
+
+  export default EntryForm;
