+++ Part2/phonebook/src/components/Person.js
@@ -1,0 +1,3 @@
+const Person = ({name, number}) => <li>{name} {number}</li>
+
+export default Person;
